requirements.txt
Django==2.0
Jinja2==2.8
gunicorn==19.7.1
celery==4.2.0
whitenoise==3.1
django-taggit==0.22.2